+++ lib/auth_gate.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:thundercard/main.dart';
+import 'package:thundercard/home_page.dart';
 import 'package:flutterfire_ui/auth.dart';
 
 class AuthGate extends StatelessWidget {
@@ -17,7 +17,7 @@ class AuthGate extends StatelessWidget {
           ]);
         }
 
-        return MyHomePage();
+        return HomePage();
       },
     );
   }
